Connected user with books added from him

+++ src/components/My-books/my-books.component.jsx
@@ -9,23 +9,27 @@ import './my-books.styles.scss'
 class MyBooks extends Component {
 
   render() {
-    const { books } = this.props;
+    const { books, auth } = this.props;
     const books_th = books.map(book => {
-      return (
-        <tr key={book.id}>
-          <th><Link to={`/my-books/${book.id}`}>{book.title}</Link></th>
-          <th>{book.author}</th>
-          <th>{book.note}</th>
-          <th>{book.state}</th>
-          <th>{book.rate}</th>
-          <th className="actions-row" id={book.id}>
-            <button onClick={() => {
-              deleteConfirm(book.id)
-            }}><i className="far fa-trash-alt"></i></button>
-            <button><i className="far fa-edit"></i></button>
-          </th>
-        </tr>
-      )
+      if (book.userID === auth.uid) {
+        return (
+          <tr key={book.id}>
+            <th><Link to={`/my-books/${book.id}`}>{book.title}</Link></th>
+            <th>{book.author}</th>
+            <th>{book.note}</th>
+            <th>{book.state}</th>
+            <th>{book.rate}</th>
+            <th className="actions-row" id={book.id}>
+              <button onClick={() => {
+                deleteConfirm(book.id)
+              }}><i className="far fa-trash-alt"></i></button>
+              <button><i className="far fa-edit"></i></button>
+            </th>
+          </tr>
+        )
+      } else {
+        return null;
+      };
     })
 
     const deleteConfirm = (bookID) => {
@@ -58,7 +62,8 @@ class MyBooks extends Component {
 
 const mapStateToProps = (state) => {
   return {
-    books: state.firestore.ordered.books || state.books.books
+    books: state.firestore.ordered.books || state.books.books,
+    auth: state.firebase.auth
   }
 }
